requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
scikit-learn==1.2.2
streamlit==1.16.0
xgboost==2.1.4
